pyproject.toml
[tool.poetry]
name = "foothold-sitac"
version = "0.1.0"
description = ""
authors = ["Michel NAUD <michel.naud26400@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.121.1"
uvicorn = "^0.38.0"
jinja2 = "^3.1.6"
pydantic = "^2.12.4"
lupa = "^2.6"
pyyaml = "^6.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
